Projeto 2 - Wide-column Store/requirements.txt
cassandra-driver
Faker